remove support for int, int64 is more useful

+++ pb_test.go
@@ -5,9 +5,9 @@ import (
 )
 
 func Test_IncrementAddsOne(t *testing.T) {
-	count := 5000
+	count := int64(5000)
 	bar := New(count)
-	expected := 1
+	expected := int64(1)
 	actual := bar.Increment()
 
 	if actual != expected {
@@ -16,7 +16,7 @@ func Test_IncrementAddsOne(t *testing.T) {
 }
 
 func Test_Width(t *testing.T) {
-	count := 5000
+	count := int64(5000)
 	bar := New(count)
 	width := 100
 	bar.SetWidth(100).Callback = func(out string) {
